+++ syndicode-server/src/application/ports/processor.rs
@@ -4,5 +4,5 @@ use mockall::{automock, predicate::*};
 #[cfg_attr(test, automock)]
 #[tonic::async_trait]
 pub trait GameTickProcessable: Send + Sync {
-    async fn process_next_tick(&self) -> anyhow::Result<usize>;
+    async fn process_next_tick(&self) -> anyhow::Result<i64>;
 }
